// client-pokemon/src/interfaces/interfaces.tsx
// Generated by https://quicktype.io

export interface Pokemons {
    data: Pokemon[];
}

export interface Pokemon {
    name:     string;
    height:   number;
    weight:   number;
    id:       number;
    location: string;
    moves:    Move[];
    types:    Type[];
}

export interface Move {
    name:          string;
    pokemon_moves: PokemonMoves;
}

export interface PokemonMoves {
    moveName: string;
}

export interface Type {
    name:          string;
    pokemon_types: PokemonTypes;
}

export interface PokemonTypes {
    typeName: string;
}